make sure placeholder file in uploads folder does not get deleted, so that the folder persists

+++ handlers/fileSystem.js
@@ -2,23 +2,46 @@ const path = require("path");
 const fs = require("fs");
 const directory = "public/uploads";
 
+
+
 const goOnAndDelete = () => {
   fs.readdir(directory, (err, files) => {
     if (err) console.error(err);
-    
-    console.log(`There are ${files.length} pictures in the uploads folder`)
-    
+
     if (files.length === 0) {
       return;
     }
 
     for (const file of files) {
-      fs.unlink(path.join(directory, file), err => {
-        if (err) throw err;
-      });
+      const fileDir = path.join("./", file);
+      console.log("fileDir", fileDir);
+      if (fileDir !== "placeholder.js") {
+        fs.unlink(path.join(directory, file), err => {
+          if (err) throw err;
+        });
+      }
     }
   });
-}
+};
+
+
+// const goOnAndDelete = () => {
+//   fs.readdir(directory, (err, files) => {
+//     if (err) console.error(err);
+    
+//     console.log(`There are ${files.length} pictures in the uploads folder`)
+    
+//     if (files.length === 0) {
+//       return;
+//     }
+
+//     for (const file of files) {
+//       fs.unlink(path.join(directory, file), err => {
+//         if (err) throw err;
+//       });
+//     }
+//   });
+// }
 
 exports.deleteAllImagesInUploadsFolder = () => {
   //check if folder exists
